refactor(studio-chat): split message store and streaming runtime
