+++ packages/studio/src/store/chat/slices/message/runtime-labels.ts
@@ -1,0 +1,19 @@
+const AGENT_LABELS: Record<string, string> = {
+  architect: "建书",
+  writer: "写作",
+  auditor: "审计",
+  reviser: "修订",
+  exporter: "导出",
+};
+
+const TOOL_LABELS: Record<string, string> = {
+  read: "读取文件",
+  edit: "编辑文件",
+  grep: "搜索",
+  ls: "列目录",
+};
+
+export function resolveToolLabel(tool: string, agent?: string): string {
+  if (tool === "sub_agent" && agent) return AGENT_LABELS[agent] ?? agent;
+  return TOOL_LABELS[tool] ?? tool;
+}
