+++ packages/studio/src/store/chat/slices/message/runtime-parts.ts
@@ -1,0 +1,95 @@
+import type {
+  Message,
+  MessagePart,
+  SessionMessage,
+  ToolExecution,
+} from "../../types";
+
+export function getOrCreateStream(
+  messages: ReadonlyArray<Message>,
+  streamTs: number,
+): [ReadonlyArray<Message>, Message] {
+  const last = messages[messages.length - 1];
+  if (last?.timestamp === streamTs && last.role === "assistant") {
+    return [messages, last];
+  }
+  const message: Message = { role: "assistant", content: "", timestamp: streamTs, parts: [] };
+  return [[...messages, message], message];
+}
+
+export function replaceLast(
+  messages: ReadonlyArray<Message>,
+  updated: Message,
+): ReadonlyArray<Message> {
+  return [...messages.slice(0, -1), updated];
+}
+
+export function findRunningToolPart(
+  parts: MessagePart[],
+): (MessagePart & { type: "tool" }) | undefined {
+  for (let i = parts.length - 1; i >= 0; i -= 1) {
+    const part = parts[i];
+    if (part.type === "tool" && part.execution.status === "running") {
+      return part as MessagePart & { type: "tool" };
+    }
+  }
+  return undefined;
+}
+
+export function deriveFlat(
+  parts: MessagePart[],
+): { content: string; thinking?: string; thinkingStreaming?: boolean; toolExecutions?: ToolExecution[] } {
+  let content = "";
+  let thinking = "";
+  let thinkingStreaming = false;
+  const toolExecutions: ToolExecution[] = [];
+
+  for (const part of parts) {
+    if (part.type === "thinking") {
+      if (thinking) thinking += "\n\n---\n\n";
+      thinking += part.content;
+      if (part.streaming) thinkingStreaming = true;
+      continue;
+    }
+
+    if (part.type === "text") {
+      content += part.content;
+      continue;
+    }
+
+    toolExecutions.push(part.execution);
+  }
+
+  return {
+    content,
+    thinking: thinking || undefined,
+    thinkingStreaming: thinkingStreaming ? true : undefined,
+    toolExecutions: toolExecutions.length > 0 ? toolExecutions : undefined,
+  };
+}
+
+export function deserializeMessages(
+  msgs: ReadonlyArray<SessionMessage>,
+): ReadonlyArray<Message> {
+  return msgs
+    .filter((message) => message.role === "user" || message.role === "assistant")
+    .map((message) => {
+      const toolExecutions = (message as any).toolExecutions as ToolExecution[] | undefined;
+      const parts: MessagePart[] = [];
+      if (message.thinking) parts.push({ type: "thinking", content: message.thinking, streaming: false });
+      if (toolExecutions) {
+        for (const execution of toolExecutions) {
+          parts.push({ type: "tool", execution });
+        }
+      }
+      if (message.content) parts.push({ type: "text", content: message.content });
+      return {
+        role: message.role as "user" | "assistant",
+        content: message.content,
+        thinking: message.thinking,
+        toolExecutions,
+        timestamp: message.timestamp,
+        parts: parts.length > 0 ? parts : undefined,
+      };
+    });
+}
