+++ packages/studio/src/store/chat/slices/message/runtime-sessions.ts
@@ -1,0 +1,75 @@
+import type {
+  Message,
+  SessionRuntime,
+  SessionSummary,
+} from "../../types";
+
+const NULL_BOOK_KEY = "__null__";
+
+export function bookKey(bookId: string | null | undefined): string {
+  return bookId ?? NULL_BOOK_KEY;
+}
+
+export function createSessionRuntime(input: {
+  sessionId: string;
+  bookId: string | null;
+  title: string | null;
+  messages?: ReadonlyArray<Message>;
+  isDraft?: boolean;
+}): SessionRuntime {
+  return {
+    sessionId: input.sessionId,
+    bookId: input.bookId,
+    title: input.title,
+    messages: input.messages ?? [],
+    stream: null,
+    isStreaming: false,
+    lastError: null,
+    isDraft: input.isDraft ?? false,
+  };
+}
+
+export function updateSession(
+  sessions: Record<string, SessionRuntime>,
+  sessionId: string,
+  updater: (session: SessionRuntime) => Partial<SessionRuntime>,
+): Record<string, SessionRuntime> {
+  const existing = sessions[sessionId];
+  if (!existing) return sessions;
+  return {
+    ...sessions,
+    [sessionId]: {
+      ...existing,
+      ...updater(existing),
+    },
+  };
+}
+
+export function upsertSessionSummary(
+  sessions: Record<string, SessionRuntime>,
+  summary: Pick<SessionSummary, "sessionId" | "bookId" | "title">,
+): Record<string, SessionRuntime> {
+  const existing = sessions[summary.sessionId];
+  return {
+    ...sessions,
+    [summary.sessionId]: existing
+      ? { ...existing, bookId: summary.bookId, title: summary.title }
+      : createSessionRuntime(summary),
+  };
+}
+
+export function mergeSessionIds(
+  existing: ReadonlyArray<string> | undefined,
+  incoming: ReadonlyArray<string>,
+): ReadonlyArray<string> {
+  if (!existing?.length) return [...incoming];
+  const seen = new Set(existing);
+  const appended = incoming.filter((id) => !seen.has(id));
+  if (appended.length === 0) return existing as string[];
+  return [...existing, ...appended];
+}
+
+export function sessionMatchesEvent(sessionId: string, data: unknown): boolean {
+  if (!data || typeof data !== "object") return false;
+  return (data as { sessionId?: unknown }).sessionId === sessionId;
+}
